ContractAI-backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.5
minio==7.1.12
pytesseract==0.3.9
pdf2image==1.16.3
Pillow==9.5.0
python-dotenv==0.21.0
sqlalchemy==1.4.47
psycopg2-binary==2.9.6
